requirements.txt
alembic==1.0.8
bcrypt==3.1.6
cffi==1.12.2
Click==7.0
contextlib2==0.5.5
Flask==1.0.2
Flask-Bcrypt==0.7.1
flask-marshmallow==0.10.0
Flask-Migrate==2.4.0
Flask-SQLAlchemy==2.3.2
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.8
MarkupSafe==1.1.1
marshmallow==2.19.2
marshmallow-sqlalchemy==0.16.1
psycopg2==2.8.1
pycparser==2.19
python-dateutil==2.8.0
python-dotenv==0.10.1
python-editor==1.0.4
schema==0.7.0
six==1.12.0
SQLAlchemy==1.3.2
Werkzeug==0.15.2
